Moved all plugin classes into plugins subfolder
