+++ fastlane/lib/fastlane.rb
@@ -18,7 +18,7 @@ require 'fastlane/one_off'
 require 'fastlane/command_line_handler'
 require 'fastlane/documentation/docs_generator'
 require 'fastlane/other_action'
-require 'fastlane/plugin_manager'
+require 'fastlane/plugins/plugin_manager'
 
 module Fastlane
   Helper = FastlaneCore::Helper # you gotta love Ruby: Helper.* should use the Helper class contained in FastlaneCore
