+++ fastlane/lib/fastlane/plugins/plugin_manager.rb
